Move help modal and buttons to parent wrapper
