+++ client/src/components/Markdown/Editor/TextEditor.vue
@@ -5,56 +5,23 @@
         </FlexPanel>
         <div id="center" class="overflow-auto w-100">
             <div class="markdown-editor h-100">
-                <div class="unified-panel-header" unselectable="on">
-                    <div class="unified-panel-header-inner">
-                        <div class="panel-header-buttons">
-                            <slot name="buttons" />
-                            <b-button
-                                v-b-tooltip.hover.bottom
-                                title="Help"
-                                variant="link"
-                                role="button"
-                                @click="onHelp">
-                                <FontAwesomeIcon icon="question" />
-                            </b-button>
-                        </div>
-                        <div class="my-1">
-                            {{ title }}
-                        </div>
-                    </div>
-                </div>
-                <div class="unified-panel-body d-flex">
-                    <textarea
-                        id="workflow-report-editor"
-                        ref="textArea"
-                        v-model="content"
-                        class="markdown-textarea"
-                        @input="onUpdate" />
-                </div>
+                <textarea
+                    id="workflow-report-editor"
+                    ref="textArea"
+                    v-model="content"
+                    class="markdown-textarea"
+                    @input="onUpdate" />
             </div>
         </div>
-        <b-modal v-model="showHelpModal" hide-footer>
-            <template v-slot:modal-title>
-                <h2 v-if="mode === 'page'" class="mb-0">Markdown Help for Pages</h2>
-                <h2 v-else class="mb-0">Markdown Help for Invocation Reports</h2>
-            </template>
-            <MarkdownHelp :mode="mode" />
-        </b-modal>
     </div>
 </template>
 
 <script setup lang="ts">
-import { library } from "@fortawesome/fontawesome-svg-core";
-import { faQuestion } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 import { debounce } from "lodash";
 import { defineEmits, defineProps, nextTick, ref, watch } from "vue";
 
-import MarkdownHelp from "@/components/Markdown/MarkdownHelp.vue";
 import MarkdownToolBox from "@/components/Markdown/MarkdownToolBox.vue";
 import FlexPanel from "@/components/Panels/FlexPanel.vue";
-
-library.add(faQuestion);
 
 const props = defineProps<{
     markdownText: string;
@@ -68,7 +35,6 @@ const emit = defineEmits<{
 }>();
 
 const content = ref<string>(props.markdownText);
-const showHelpModal = ref<boolean>(false);
 const textArea = ref<HTMLTextAreaElement | null>(null);
 
 const FENCE = "```";
@@ -105,8 +71,4 @@ function insertMarkdown(markdown: string) {
 const onUpdate = debounce((e: Event) => {
     emit("update", content.value || "");
 }, 300);
-
-function onHelp() {
-    showHelpModal.value = true;
-}
 </script>
